Remove outdated documentation files for media and ticket comments features
This commit deletes several markdown files that contained implementation summaries and user guides for the "Add Manual Action" and "Ticket Comments" features, as well as the "Media API Guide" and "Uploader and Media Manager" documentation. These files are no longer needed as the features have been fully integrated and documented elsewhere in the project. This cleanup helps maintain a more organized codebase.